requirements.txt
Django==4.1.3
python-dotenv==0.21.0
pika==1.3.1
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
environs==9.5.0

